// server/graphql/types/v2/schema.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  /** Values that can be edited in Account's settings */
  AccountSettingsKey: { input: any; output: any };
  /** A date string, such as 2007-12-03, compliant with the `full-date` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  Date: { input: any; output: any };
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: { input: any; output: any };
  /** A field whose value conforms to the standard internet email address format as specified in RFC822: https://www.w3.org/Protocols/rfc822/. */
  EmailAddress: { input: any; output: any };
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any };
  /** The `JSONObject` scalar type represents JSON objects as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSONObject: { input: any; output: any };
  /** The locale in the format of a BCP 47 (RFC 5646) standard string */
  Locale: { input: any; output: any };
  /** A string that cannot be passed as an empty value */
  NonEmptyString: { input: any; output: any };
  /** A positive float value between 0 and 100 */
  StrictPercentage: { input: any; output: any };
  /** A field whose value conforms to the standard URL format as specified in RFC3986: https://www.ietf.org/rfc/rfc3986.txt. */
  URL: { input: any; output: any };
  /** The `Upload` scalar type represents a file upload. */
  Upload: { input: any; output: any };
};
// ...
export type Amount = {
  __typename?: 'Amount';
  currency?: Maybe<Currency>;
  /** If the amount was generated from a currency conversion, this field contains details about the conversion */
  exchangeRate?: Maybe<CurrencyExchangeRate>;
  value?: Maybe<Scalars['Float']['output']>;
  valueInCents?: Maybe<Scalars['Float']['output']>;
};
// ...
export type Order = {
  __typename?: 'Order';
  /** The accounting category attached to this order */
  accountingCategory?: Maybe<AccountingCategory>;
  /** The list of activities (ie. approved, edited, etc) for this Order ordered by date ascending */
  activities: ActivityCollection;
  /** Base order amount (without platform tip) */
  amount: Amount;
  /** Returns the list of comments for this order, or `null` if user is not allowed to see them */
  comments?: Maybe<CommentCollection>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  /** The account who created this order */
  createdByAccount?: Maybe<Account>;
  /** Custom data related to the order, based on the fields described by tier.customFields. Must be authenticated as an admin of the fromAccount or toAccount (returns null otherwise) */
  customData?: Maybe<Scalars['JSON']['output']>;
  /** Data related to the order */
  data?: Maybe<Scalars['JSON']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  frequency?: Maybe<ContributionFrequency>;
  fromAccount?: Maybe<Account>;
  /** Host fee percent attached to the Order. */
  hostFeePercent?: Maybe<Scalars['Float']['output']>;
  id: Scalars['String']['output'];
  lastChargedAt?: Maybe<Scalars['DateTime']['output']>;
  legacyId: Scalars['Int']['output'];
  /** This represents a MemberOf relationship (ie: Collective backed by an Individual) attached to the Order. */
  membership?: Maybe<MemberOf>;
  /** Memo field which adds additional details about the order. For example in added funds this can be a note to mark what method (cheque, money order) the funds were received. */
  memo?: Maybe<Scalars['String']['output']>;
  /** Whether the order needs confirmation (3DSecure/SCA) */
  needsConfirmation?: Maybe<Scalars['Boolean']['output']>;
  nextChargeDate?: Maybe<Scalars['DateTime']['output']>;
  paymentMethod?: Maybe<PaymentMethod>;
  /** Payment processor fee attached to manually Added Funds. */
  paymentProcessorFee?: Maybe<Amount>;
  paymentProcessorUrl?: Maybe<Scalars['String']['output']>;
  /** Data about the pending contribution */
  pendingContributionData?: Maybe<PendingOrderData>;
  /** The permissions given to current logged in user for this order */
  permissions: OrderPermissions;
  /** Platform Tip attached to the Order. */
  platformTipAmount?: Maybe<Amount>;
  platformTipEligible?: Maybe<Scalars['Boolean']['output']>;
  /** Date the funds were received. */
  processedAt?: Maybe<Scalars['DateTime']['output']>;
  quantity?: Maybe<Scalars['Int']['output']>;
  status?: Maybe<OrderStatus>;
  tags: Array<Maybe<Scalars['String']['output']>>;
  tax?: Maybe<TaxInfo>;
  /** Tax amount */
  taxAmount?: Maybe<Amount>;
  /** @deprecated 2023-04-13: Please use `tax` instead. */
  taxes: Array<Maybe<OrderTax>>;
  tier?: Maybe<Tier>;
  toAccount?: Maybe<Account>;
  /** Total order amount, including all taxes and platform tip */
  totalAmount: Amount;
  /** Total amount contributed with this order. */
  totalContributed: Amount;
  /** WARNING: Total amount donated between collectives, though there will be edge cases especially when looking on the Order level, as the order id is not used in calculating this. */
  totalDonations: Amount;
  /** [Host admins only] If the order was associated with a transactions import row, this field will reference it */
  transactionImportRow?: Maybe<TransactionsImportRow>;
  /** Transactions for this order ordered by createdAt ASC */
  transactions: Array<Maybe<Transaction>>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};
// ...
export type PaymentMethod = {
  __typename?: 'PaymentMethod';
  account?: Maybe<Account>;
  /** Returns the balance amount and the currency of this paymentMethod */
  balance: Amount;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  data?: Maybe<Scalars['JSON']['output']>;
  expiryDate?: Maybe<Scalars['DateTime']['output']>;
  id?: Maybe<Scalars['String']['output']>;
  legacyId?: Maybe<Scalars['Int']['output']>;
  limitedToHosts?: Maybe<Array<Maybe<Host>>>;
  /** For monthly gift cards, this field will return the monthly limit */
  monthlyLimit?: Maybe<Amount>;
  name?: Maybe<Scalars['String']['output']>;
  /** Get all the orders associated with this payment method */
  orders?: Maybe<OrderCollection>;
  /**
   * Defines the type of the payment method. Meant to be moved to "type" in the future.
   * @deprecated 2021-03-02: Please use service + type
   */
  providerType?: Maybe<PaymentMethodLegacyType>;
  service?: Maybe<PaymentMethodService>;
  /** For gift cards, this field will return to the source payment method */
  sourcePaymentMethod?: Maybe<PaymentMethod>;
  type?: Maybe<PaymentMethodType>;
};
// ...
export type PayoutMethod = {
  __typename?: 'PayoutMethod';
  /** Whether this payout method can be edit or deleted */
  canBeEditedOrDeleted?: Maybe<Scalars['Boolean']['output']>;
  /** The actual data for this payout method. Content depends on the type. */
  data?: Maybe<Scalars['JSON']['output']>;
  /** Unique identifier for this payout method */
  id: Scalars['String']['output'];
  /** Whether this payout method has been saved to be used for future payouts */
  isSaved?: Maybe<Scalars['Boolean']['output']>;
  /** A friendly name for users to easily find their payout methods */
  name?: Maybe<Scalars['String']['output']>;
  /** The type of this payout method (usually the payment provider) */
  type?: Maybe<PayoutMethodType>;
};
// ...
export type TaxInfo = {
  __typename?: 'TaxInfo';
  /** An unique identifier for this tax (GST, VAT, etc) */
  id: Scalars['String']['output'];
  /** Tax ID number of the 3rd party receiving/paying the tax */
  idNumber?: Maybe<Scalars['String']['output']>;
  /**
   * Percentage applied, between 0-100
   * @deprecated Please use `rate` instead
   */
  percentage: Scalars['Int']['output'];
  /** Percentage applied, between 0-1 */
  rate: Scalars['Float']['output'];
  /** Identifier for this tax (GST, VAT, etc) */
  type: TaxType;
};
// ...
export type Transaction = {
  account?: Maybe<Account>;
  amount: Amount;
  amountInHostCurrency: Amount;
  /** The balance after the Transaction has run. Only for financially active accounts. */
  balanceInHostCurrency?: Maybe<Amount>;
  clearedAt?: Maybe<Scalars['DateTime']['output']>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  expense?: Maybe<Expense>;
  /** The sender of a transaction (on CREDIT = oppositeAccount, DEBIT = account) */
  fromAccount?: Maybe<Account>;
  giftCardEmitterAccount?: Maybe<Account>;
  group: Scalars['String']['output'];
  host?: Maybe<Account>;
  /** Exchange rate between the currency of the transaction and the currency of the host (transaction.amount * transaction.hostCurrencyFxRate = transaction.amountInHostCurrency) */
  hostCurrencyFxRate?: Maybe<Scalars['Float']['output']>;
  hostFee?: Maybe<Amount>;
  id: Scalars['String']['output'];
  invoiceTemplate?: Maybe<Scalars['String']['output']>;
  isDisputed?: Maybe<Scalars['Boolean']['output']>;
  isInReview?: Maybe<Scalars['Boolean']['output']>;
  isOrderRejected: Scalars['Boolean']['output'];
  isRefund?: Maybe<Scalars['Boolean']['output']>;
  isRefunded?: Maybe<Scalars['Boolean']['output']>;
  kind?: Maybe<TransactionKind>;
  legacyId: Scalars['Int']['output'];
  /** Merchant id related to the Transaction (Stripe, PayPal, Wise, Privacy) */
  merchantId?: Maybe<Scalars['String']['output']>;
  netAmount: Amount;
  netAmountInHostCurrency: Amount;
  oppositeAccount?: Maybe<Account>;
  /** The opposite transaction (CREDIT -> DEBIT, DEBIT -> CREDIT) */
  oppositeTransaction?: Maybe<Transaction>;
  order?: Maybe<Order>;
  paymentMethod?: Maybe<PaymentMethod>;
  paymentProcessorFee?: Maybe<Amount>;
  paymentProcessorUrl?: Maybe<Scalars['String']['output']>;
  payoutMethod?: Maybe<PayoutMethod>;
  permissions?: Maybe<TransactionPermissions>;
  platformFee: Amount;
  refundTransaction?: Maybe<Transaction>;
  relatedTransactions: Array<Maybe<Transaction>>;
  taxAmount: Amount;
  /** If a tax is set, this field will contain more info about the tax */
  taxInfo?: Maybe<TaxInfo>;
  /** The recipient of a transaction (on CREDIT = account, DEBIT = oppositeAccount) */
  toAccount?: Maybe<Account>;
  type: TransactionType;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  /** @deprecated 2021-08-15: Use id instead. */
  uuid: Scalars['String']['output'];
};
// ...
export enum TransactionKind {
  ADDED_FUNDS = 'ADDED_FUNDS',
  BALANCE_TRANSFER = 'BALANCE_TRANSFER',
  CONTRIBUTION = 'CONTRIBUTION',
  EXPENSE = 'EXPENSE',
  HOST_FEE = 'HOST_FEE',
  HOST_FEE_SHARE = 'HOST_FEE_SHARE',
  HOST_FEE_SHARE_DEBT = 'HOST_FEE_SHARE_DEBT',
  PAYMENT_PROCESSOR_COVER = 'PAYMENT_PROCESSOR_COVER',
  PAYMENT_PROCESSOR_DISPUTE_FEE = 'PAYMENT_PROCESSOR_DISPUTE_FEE',
  PAYMENT_PROCESSOR_FEE = 'PAYMENT_PROCESSOR_FEE',
  PLATFORM_FEE = 'PLATFORM_FEE',
  PLATFORM_TIP = 'PLATFORM_TIP',
  PLATFORM_TIP_DEBT = 'PLATFORM_TIP_DEBT',
  PREPAID_PAYMENT_METHOD = 'PREPAID_PAYMENT_METHOD',
  TAX = 'TAX',
}

/** Fields for the user permissions on an transaction */
export type TransactionPermissions = {
  __typename?: 'TransactionPermissions';
  /** Whether the current user can download this transaction's invoice */
  canDownloadInvoice: Scalars['Boolean']['output'];
  /** Whether the current user can edit the transaction */
  canRefund: Scalars['Boolean']['output'];
  /** Whether the current user can reject the transaction */
  canReject: Scalars['Boolean']['output'];
  id: Scalars['String']['output'];
};
// ...
export enum TransactionType {
  CREDIT = 'CREDIT',
  DEBIT = 'DEBIT',
}